+++ apps/web/app/app.dub.co/(dashboard)/[slug]/(ee)/program/analytics/[tab]/page.tsx
@@ -1,0 +1,1 @@
+export { default } from "../page";
